requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
argparse==1.4.0           # via jira
arrow==0.10.0
asn1crypto==0.22.0        # via cryptography
astroid==1.5.2            # via edx-lint, pylint, pylint-celery, pylint-plugin-utils
backports.functools-lru-cache==1.4
betamax==0.8.0
billiard==3.3.0.23        # via celery
blinker==1.4
celery==3.1.25
certifi==2017.7.27.1      # via requests
cffi==1.10.0              # via cryptography
chardet==3.0.4            # via requests
click-log==0.1.8          # via edx-lint
click==6.7
codecov==2.0.9
configparser==3.5.0       # via pylint
contextlib2==0.5.5        # via raven
coverage==4.4.1           # via codecov, pytest-cov
cryptography==2.0.3
defusedxml==0.5.0         # via jira
edx-lint==0.5.4
enum34==1.1.6             # via astroid, cryptography
first==2.0.1              # via pip-tools
flask-dance[sqla]==0.9.0
flask-script==2.0.6
flask-sqlalchemy==2.0
flask-sslify==0.1.5
flask==0.10.1             # via flask-dance, flask-script, flask-sqlalchemy, flask-sslify, rq-dashboard
funcsigs==1.0.2           # via mock
functools32==3.2.3.post2
github3.py==0.9.6
gunicorn==19.7.1
idna==2.6                 # via cryptography, requests
ipaddress==1.0.18         # via cryptography
iron-cache==0.3.2         # via iron-celery
iron-celery==0.4.5
iron-core==1.2.0          # via iron-cache, iron-mq
iron-mq==0.9              # via iron-celery
iso8601==0.1.12
isort==4.2.15
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
jira==1.0.10
kombu==3.0.37             # via celery
lazy-object-proxy==1.3.1  # via astroid
lazy==1.3                 # via flask-dance
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via pylint
mock==2.0.0               # via pytest-mock
oauthlib[signedtoken]==2.0.3
ordereddict==1.1          # via jira
pbr==3.1.1                # via jira, mock
pip-tools==1.9.0
psycopg2==2.7.3.1
py==1.4.34                # via pytest
pycparser==2.18           # via cffi
pyjwt==1.5.3              # via oauthlib
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.2.6  # via pylint-celery, pylint-django
pylint==1.7.1             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pytest-cov==2.5.1
pytest-mock==1.6.2
pytest==3.2.2
python-dateutil==2.6.1    # via arrow, iron-core
python-dotenv==0.7.1
pytz==2017.2
pyyaml==3.12
raven==6.1.0
redis==2.10.6
requests-mock==1.3.0
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0  # via jira
requests==2.18.4
rq-dashboard==0.3.8
rq==0.8.2
singledispatch==3.4.0.3   # via astroid, pylint
six==1.10.0               # via astroid, cryptography, edx-lint, flask-dance, jira, mock, pip-tools, pylint, python-dateutil, requests-mock, singledispatch, sqlalchemy-utils
sqlalchemy-utils==0.32.16  # via flask-dance
sqlalchemy==1.1.14        # via flask-dance, flask-sqlalchemy, sqlalchemy-utils
uritemplate.py==3.0.2     # via github3.py
uritemplate==3.0.0        # via uritemplate.py
urllib3==1.22             # via requests
urlobject==2.4.3          # via flask-dance
werkzeug==0.12.2          # via flask
wrapt==1.10.11            # via astroid
